host: Move ps, stop and log commands into cli package
